Capture the stderr output from Git when collecting progress
The design of the Process API in ToolSupportCore is really suboptimal.  It should be possible to watch an output stream without blocking collection of that stream.  A better design would allow zero or more callbacks for each of the output streams (stdout and stderr), and with an option to also collect each one (independently of whether they're being watched).  Or possibly allow a filter chain, etc.  In any case, this works around it by capturing stderr.  We change the GitCloneError to just store the output of the ProcessResult, which is all that was being used.  If we need the rest of the ProcessResult information in the future (which we probably should use), we can add it.  But probably fix TSC first so we don't need this.