[WIP] Extend the plugin API by passing down a lot more information about the package graph
The original TargetBuildContext was focused on build tools, which was the subject of SE-0303.  In order to prepare for new kinds of plugins such as custom commands (which have much more varied purposes and need more diverse information), we generalize this structure into a PluginContext that contains information about a whole package graph.

The graph (being a DAG) is flattened and serialized into the plugin input JSON, and is reconstituted on the plugin side.  Path deduplication is used to reduce the amount of information sent over the wire.

A compatibility layer lets old plugins keep working with the old TargetBuildContext.

This is a work in progress.  There are some unaddressed issues labeled in the code, and unit tests for the new APIs are still missing.  The existing plugin test fixtures still use the old TargetBuildContext API, but they are working, due to the compatibility layer.