[5.5] Loosen the assumptions around product types, letting clients of libSwiftPM provide additional ones
This starts to unwind some of the hardcoding around product types, and lets libSwiftPM clients provide additional product types through add-on libraries loaded on top of `PackageDescription`.

Note that this isn’t yet plugin-provided product types (that will be a much larger effort).  But it is a small step towards it.

The existing product types of executable, library, etc are still kept throughout libSwiftPM, but the way in which they are encoded in PackageDescription and read by libSwiftPM is now more generic.  Additionally, other types with custom properties are supported.  The custom properties are kept in encoded form so that libSwiftPM doesn’t need to understand the details of them.  They can be defined and encoded by additional libraries loaded by the manifest on top of PackageDescription, and clients of libSwiftPM can unpack them using the same structure definitions as in that library.  libSwiftPM of course unpacks the properties of its built-in types (currently only whether a library is static, dynamic, or automatic).

This also updates the manifest source generation to allow clients to generate source code fragments for custom product types.  A future addition would be to automatically generate these source code fragments using Mirrors.