[5.5] Additional flexibility in the manifest source generation
Allow clients of the code that can generate manifest source generation to provide a tools version header comment, a list of imports, and custom handling for product types (falling back on default behavior if they don't provide a custom source code fragment for the product).  Also makes SourceCodeFragment available to clients and adds documentation code for it.